requirements.txt
# Core PDF Processing
pdfplumber>=0.10.0

# OCR for scanned documents
ocrmypdf>=15.0.0

# Data validation and parsing
pydantic>=2.0.0
python-dateutil>=2.8.0

# Configuration management
PyYAML>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
